+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Next Up
 
 - Improve Compendium Folder generation speed for Items and Spells.
+- Some fixes for Homebrew Monster import: @baiseralice
+- Some internal adjustments for Foundry v13 (DDB Importer is not yet supported due to some outstanding compendium bugs).
 
 # 6.0.78